+++ tests/conftest.py
@@ -1,0 +1,17 @@
+list_double_quotes = '["ABC", "123", "Pau", "F12"]'
+list_single_quotes = "['ABC', '123', 'Pau', 'F12']"
+
+default_category = [
+    "BG",
+    "CH",
+    "FX",
+    "Props",
+    "Vehicles",
+]
+
+default_subfolder = [
+    "References",
+    "Scenes",
+    "Textures",
+    "WIP",
+]